Add new version

+++ src/Glpi/Asset/Capacity/HasDevicesCapacity.php
@@ -36,6 +36,7 @@ namespace Glpi\Asset\Capacity;
 
 use CommonDevice;
 use CommonGLPI;
+use DBmysql;
 use Glpi\Asset\CapacityConfig;
 use Item_Devices;
 use Override;
@@ -89,7 +90,7 @@ class HasDevicesCapacity extends AbstractCapacity
     private function countAssetsLinkedToDevices(string $classname): int
     {
         /**
-         * @var \DBmysql $DB
+         * @var DBmysql $DB
          */
         global $DB;
 
@@ -157,6 +158,10 @@ class HasDevicesCapacity extends AbstractCapacity
         global $CFG_GLPI;
 
         foreach (Item_Devices::getDeviceTypes() as $item_device_class) {
+            if (!\is_a($item_device_class, Item_Devices::class, true)) {
+                continue;
+            }
+
             //Delete related items
             $item_device = new $item_device_class();
             $item_device->deleteByCriteria(['itemtype' => $classname], force: true, history: false);
